Move TableGenerator to benchmarklib.
