+++ src/benchmark/CMakeLists.txt
@@ -17,8 +17,6 @@ add_executable(
     operators/sql_benchmark.cpp
     operators/table_scan_benchmark.cpp
     operators/union_all_benchmark.cpp
-    table_generator.cpp
-    table_generator.hpp
     tpch_db_generator_benchmark.cpp
 )
