+++ src/benchmark/operators/join_benchmark.cpp
@@ -1,7 +1,7 @@
 #include <memory>
 
 #include "../benchmark_join_fixture.hpp"
-#include "../table_generator.hpp"
+#include "table_generator.hpp"
 #include "benchmark/benchmark.h"
 #include "operators/join_hash.hpp"
 #include "operators/join_index.hpp"
