+++ src/benchmark/operators/projection_benchmark.cpp
@@ -4,7 +4,7 @@
 #include "benchmark/benchmark.h"
 
 #include "../benchmark_basic_fixture.hpp"
-#include "../table_generator.hpp"
+#include "table_generator.hpp"
 #include "operators/pqp_expression.hpp"
 #include "operators/projection.hpp"
 #include "operators/table_scan.hpp"
