+++ src/benchmark/operators/union_all_benchmark.cpp
@@ -1,7 +1,7 @@
 #include <memory>
 
 #include "../benchmark_basic_fixture.hpp"
-#include "../table_generator.hpp"
+#include "table_generator.hpp"
 #include "benchmark/benchmark.h"
 #include "operators/table_wrapper.hpp"
 #include "operators/union_all.hpp"
