+++ src/benchmarklib/CMakeLists.txt
@@ -20,6 +20,9 @@ set(
     tpch/tpch_queries.hpp
     tpch/tpch_db_generator.cpp
     tpch/tpch_db_generator.hpp
+
+    table_generator.cpp
+    table_generator.hpp
 )
